Update copyright
As per http://choosealicense.com/licenses/ and http://www.gnu.org/licenses/license-list.html we are "stuck" with GNU 2, as the original source code of the fork is licensed under it.

Will do a bit of more research if we can upgrade to GNU 2, but will leave this copyright for now.